# api/openapi.yaml
openapi: 3.0.3
info:
  title: AskAtlas API
  version: 1.0.0
servers:
  - url: /api
security:
  - BearerAuth: []
paths:
...
  /files/{file_id}/grants:
    post:
      operationId: CreateGrant
      summary: Grant a permission on a file (ASK-122)
      description: |
        Creates a new file_grants row scoped to (file_id, grantee_type,
        grantee_id, permission). Only the file owner may create grants.

        The grantee_id is validated against the corresponding table:
        users for grantee_type=user, courses for course, and study_guides
        (filtered by deleted_at IS NULL) for study_guide. The public
        sentinel UUID 00000000-0000-0000-0000-000000000000 is exempt
        from the users lookup when grantee_type=user (represents
        "public access").

        A duplicate grant returns 409 Conflict; this endpoint does NOT
        upsert. Permission hierarchy (delete >= share >= view) is
        enforced by the read-side queries, not here.
      parameters:
        - name: file_id
          in: path
          description: The unique UUID of the file to grant access to
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGrantRequest'
      responses:
        '201':
          description: Grant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrantResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: RevokeGrant
      summary: Revoke a permission on a file (ASK-125)
      description: |
        Removes the file_grants row matched by (file_id, grantee_type,
        grantee_id, permission). Only the file owner may revoke.
        Returns 204 when a row was deleted, 404 when no matching grant
        exists. Each (grantee_type, grantee_id, permission) tuple is a
        distinct grant -- revoking `view` does NOT cascade to `share`
        or `delete`.
      parameters:
        - name: file_id
          in: path
          description: The unique UUID of the file to revoke access from
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeGrantRequest'
      responses:
        '204':
          description: Grant revoked
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /study-guides/{study_guide_id}/grants:
    get:
      operationId: ListStudyGuideGrants
      summary: List permission grants on a study guide (ASK-211)
      description: |
        Returns every `study_guide_grants` row for the guide. Only
        the guide's creator may list grants; non-creators receive 403.
      parameters:
        - name: study_guide_id
          in: path
          description: The unique UUID of the study guide
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Grants list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListStudyGuideGrantsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: CreateStudyGuideGrant
      summary: Grant a permission on a study guide (ASK-211)
      description: |
        Creates a new `study_guide_grants` row scoped to
        `(study_guide_id, grantee_type, grantee_id, permission)`.
        Only the guide's creator may create grants. A duplicate
        grant returns 409 Conflict. `grantee_type` is limited to
        `user` or `course` -- study guides cannot grant access to
        other study guides (the table enforces this via a CHECK
        constraint).
      parameters:
        - name: study_guide_id
          in: path
          description: The unique UUID of the study guide
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyGuideCreateGrantRequest'
      responses:
        '201':
          description: Grant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGuideGrantResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: RevokeStudyGuideGrant
      summary: Revoke a permission on a study guide (ASK-211)
      description: |
        Deletes the `study_guide_grants` row matched by
        `(study_guide_id, grantee_type, grantee_id, permission)`.
        Only the guide's creator may revoke. Returns 204 on success,
        404 when no matching grant exists (not idempotent -- mirrors
        file_grants).
      parameters:
        - name: study_guide_id
          in: path
          description: The unique UUID of the study guide
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyGuideRevokeGrantRequest'
      responses:
        '204':
          description: Grant revoked
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateGrantRequest:
      type: object
      description: Request body for creating a file grant
      properties:
        grantee_type:
          type: string
          enum: [user, course, study_guide]
        grantee_id:
          type: string
          format: uuid
        permission:
          type: string
          enum: [view, share, delete]
      required:
        - grantee_type
        - grantee_id
        - permission
    RevokeGrantRequest:
      type: object
      description: Request body for revoking a file grant
      properties:
        grantee_type:
          type: string
          enum: [user, course, study_guide]
        grantee_id:
          type: string
          format: uuid
        permission:
          type: string
          enum: [view, share, delete]
      required:
        - grantee_type
        - grantee_id
        - permission
    GrantResponse:
      type: object
      description: A file permission grant
      properties:
        id:
          type: string
          format: uuid
        file_id:
          type: string
          format: uuid
        grantee_type:
          type: string
        grantee_id:
          type: string
          format: uuid
        permission:
          type: string
        granted_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
      required:
        - id
        - file_id
        - grantee_type
        - grantee_id
        - permission
        - granted_by
        - created_at
    StudyGuideCreateGrantRequest:
      type: object
      description: |
        Request body for POST /api/study-guides/{study_guide_id}/grants
        (ASK-211). Mirrors CreateGrantRequest but constrains
        `grantee_type` to `user`/`course` since study guides cannot
        grant access to other study guides.
      properties:
        grantee_type:
          type: string
          enum: [user, course]
        grantee_id:
          type: string
          format: uuid
        permission:
          type: string
          enum: [view, share, delete]
      required:
        - grantee_type
        - grantee_id
        - permission
    StudyGuideRevokeGrantRequest:
      type: object
      description: |
        Request body for DELETE /api/study-guides/{study_guide_id}/grants
        (ASK-211).
      properties:
        grantee_type:
          type: string
          enum: [user, course]
        grantee_id:
          type: string
          format: uuid
        permission:
          type: string
          enum: [view, share, delete]
      required:
        - grantee_type
        - grantee_id
        - permission
    StudyGuideGrantResponse:
      type: object
      description: A study-guide permission grant (ASK-211)
      properties:
        id:
          type: string
          format: uuid
        study_guide_id:
          type: string
          format: uuid
        grantee_type:
          type: string
        grantee_id:
          type: string
          format: uuid
        permission:
          type: string
        granted_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
      required:
        - id
        - study_guide_id
        - grantee_type
        - grantee_id
        - permission
        - granted_by
        - created_at
    ListStudyGuideGrantsResponse:
      type: object
      description: |
        Collection of grants on a study guide. Sorted by created_at
        DESC so the newest share surfaces first.
      properties:
        grants:
          type: array
          items:
            $ref: '#/components/schemas/StudyGuideGrantResponse'
      required:
        - grants
...
    AppError:
      type: object
      description: Standardized error response structure matching application error domains
      properties:
        code:
          type: integer
        status:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties:
            type: string
      required:
        - code
        - status
        - message
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppError'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppError'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppError'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppError'
    Conflict:
      description: Conflict with current resource state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppError'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppError'